Correctly sanitize number fields

+++ modules/backend/widgets/Form.php
@@ -706,7 +706,7 @@ class Form extends WidgetBase
             $columnParts = Str::evalHtmlArray($field->columnName);
             $columnDotted = implode('.', $columnParts);
             $columnValue = array_get($data, $columnDotted, 0);
-            if ($field->type == 'number') $columnValue = (int) $columnValue;
+            if ($field->type == 'number') $columnValue = (float) $columnValue;
             array_set($data, $columnDotted, $columnValue);
         }
